fix: refactor and order the code
