+++ package.json
@@ -17,7 +17,7 @@
   },
   "devDependencies": {
     "@dcl/asset-packs": "^2.4.3",
-    "@dcl/js-runtime": "latest",
+    "@dcl/js-runtime": "7.8.10",
     "@dcl/sdk": "latest",
     "@typescript-eslint/eslint-plugin": "^6.15.0",
     "@typescript-eslint/parser": "^6.15.0",
